feat(encore): move DSL schema to src/types/encore-dsl + auto-derive defineEncore's dsl JSON Schema
Relocates server/encore/dsl/{at-expression,at-resolver,cadence,schema}.ts
to src/types/encore-dsl/. The files are pure Zod with no Node-only
deps so they're browser-bundle-safe — the move lets both the server
(validation) and the plugin tool definition (JSON Schema generation)
import the SAME schema as a value, with zero drift risk.

The motivation is to upgrade defineEncore's `dsl` parameter from a
bare `{ type: "object" }` to the full discriminated-union JSON
Schema (payment | service, with cadence/targets/steps/firingPlan
sub-schemas) — auto-derived via z.toJSONSchema(EncoreDslInput) so
the LLM sees the same field names, types, and constraints the
server enforces. Previously the eslint no-restricted-imports rule
forbade plugin code from importing the schema at value level
(**/server/** is restricted); src/types/ is the existing convention
for cross-cutting shapes both sides consume (session.ts, events.ts).

Import paths updated in server/encore/{closure,cycle,dispatch,
notifier,obligation,reconcile,tick}.ts.
